\name{opCodes}
\alias{opCodes}
\alias{opParams}
\title{
  Extract Operation Codes and Parameters
}
\description{
  Extract the operation codes and operation parameters from
  a \code{"DVI"} object, as generated by \code{\link{readDVI}}. 
}
\usage{
opCodes(dvi)
opParams(dvi)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dvi}{A \code{"DVI"} object.}
}
\value{
  \code{opCodes} returns a numeric vector of operation codes.

  \code{opParams} returns a named list of operation parameters.
}
\author{
  Paul Murrell
}
\examples{
dvi <- readDVI(system.file("DVI", "test-pdftex.dvi", package="dvi"))
opCodes(dvi)
opParams(dvi)
}
\keyword{ dplot }
